Remove Transcripts Dependency
We'll manage episodes in the server layer.
