+++ .sourcery-templates/DerivePartialIsos.stencil
@@ -1,6 +1,5 @@
 import ApplicativeRouter
 import Either
-import EpisodeTranscripts
 import Prelude
 
 {% for enum in types.implementing.DerivePartialIsos|enum %}
