+++ Package.swift
@@ -16,10 +16,6 @@ let package = Package(
     .package(url: "https://github.com/pointfreeco/swift-web.git", .revision("35421ef")),
     .package(url: "https://github.com/pointfreeco/Ccmark.git", .branch("master")),
     .package(url: "https://github.com/vapor/postgresql.git", from: "2.0.0"),
-
-    ProcessInfo.processInfo.environment["OSS"] == "1"
-      ? .package(url: "https://github.com/pointfreeco/episode-transcripts-oss.git", .revision("2a6472f"))
-      : .package(url: "https://github.com/mbrandonw/episode-transcripts.git", .revision("5d786e7"))
     ],
   targets: [
     .target(
@@ -37,7 +33,6 @@ let package = Package(
         "Css",
         "CssReset",
         "Either",
-        "EpisodeTranscripts",
         "Html",
         "HtmlCssSupport",
         "HttpPipeline",
@@ -64,7 +59,6 @@ let package = Package(
       name: "PointFreeTestSupport",
       dependencies: [
         "Either",
-        "EpisodeTranscripts",
         "PointFree",
         "Prelude",
         "SnapshotTesting"
