+++ Sources/PointFree/Emails/NewEpisodeEmail.swift
@@ -1,6 +1,5 @@
 import Css
 import Either
-import EpisodeTranscripts
 import Foundation
 import Html
 import HtmlCssSupport
