+++ Tests/PointFreeTests/EmailTests/NewEpisodeEmail.swift
@@ -1,4 +1,3 @@
-import EpisodeTranscripts
 import Html
 import HttpPipeline
 import Optics
