+++ Tests/PointFreeTests/EpisodeTests.swift
@@ -1,4 +1,3 @@
-import EpisodeTranscripts
 import HttpPipeline
 import Optics
 @testable import PointFree
